homework-3 ревью: переоформить contractor-контекст
