+++ src/contexts/useContractor/contractor.context.ts
@@ -1,0 +1,4 @@
+import {createContext} from "react";
+import {ContractorProps} from "./contractor.props";
+
+export const ContractorContext = createContext<ContractorProps | undefined>(undefined);
