+++ src/contexts/useContractor/contractor.props.ts
@@ -1,0 +1,8 @@
+import {Contractor} from "../../types";
+
+export type ContractorProps = {
+    contractors: Contractor[];
+    addContractor: (newContractor: Contractor) => void;
+    updateContractor: (updatedContractor: Contractor) => void;
+    deleteContractor: (id: string) => void;
+}
