+++ src/main.tsx
@@ -3,7 +3,7 @@ import './style.css'
 import {createRoot} from 'react-dom/client';
 import {App} from './components/app/app'
 import {StrictMode} from "react";
-import {ContractorProvider} from "./contexts/ContractorContext";
+import {ContractorProvider} from "./contexts/useContractor/contractor.provider";
 
 createRoot(document.getElementById('root')).render(
     <StrictMode>
